requirements.txt
requests==2.26.0
jsonschema==3.2.0
pytest==7.1.2
pytest-html~=3.2.0
PyYAML==5.4.1
selenium==4.3.0
webdriver_manager==3.8.4
allure-pytest==2.9.45
pytest-logger==0.5.1
cryptography==35.0.0
configparser~=5.0.2
lxml~=4.9.1
pytz~=2021.1
numpy~=1.23.4
IMAPClient~=2.2.0
pandas~=1.3.2
python-dateutil~=2.8.2
six~=1.16.0
future~=0.18.2
bs4~=0.0.1
beautifulsoup4~=4.10.0
holidays==0.11.3.1
py~=1.10.0
mysql-connector-python==8.0.27
aiohttp==3.8.3
redis~=4.1.4
phpserialize~=1.3
PyMySQL==1.0.2
zhon~=1.1.5
paramiko~=2.10.2
openpyxl~=3.0.9
hanzidentifier~=1.0.2
sshtunnel~=0.4.0
websockets~=10.3
json-extract~=1.1.2
pytest-xdist==2.5.0
jsondiff==2.0.0
pytest-assume~=2.4.3
dictdiffer~=0.9.0
PyJWT~=2.6.0